requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.14
langchain-core>=0.2.33
langgraph>=0.2.34
langchain-anthropic>=0.1.22
anthropic>=0.37.1
pydantic>=2.7.4
orjson>=3.10.0
streamlit>=1.34.0
